refactor: Thay đổi kiểu dữ liệu thuộc tính Balance từ decimal sang long trong ListBalanceAccount.

+++ NetCore.PortalAPI/ServerCore.PortalAPI/Models/ListBalanceAccount.cs
@@ -5,7 +5,7 @@
         public string Name { get; set; }
         public string Code { get; set; }
         public int WalletId { get; set; }
-        public decimal Balance { get; set; }
+        public long Balance { get; set; }
         public int DecimalDigits { get; set; }
     }
 }
